pyproject.toml
[tool.poetry]
name = "entitlements-service"
version = "0.1.0"
description = "Entitlements service for managing access to various systems."
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10" # Though user asked for 3.13, poetry might be more flexible with ^3.10
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9" # For SQLAlchemy sync, asyncpg for async
pydantic = {extras = ["email"], version = "^2.5.2"}
alembic = "^1.13.0"
python-dotenv = "^1.0.0"
asyncpg = "^0.29.0" # For FastAPI async with SQLAlchemy
greenlet = "^3.0.3" # Required by SQLAlchemy async with psycopg2

[tool.poetry.dev-dependencies]
pytest = "^7.4.3"
httpx = "^0.25.1" # For testing FastAPI async endpoints

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
